automacao-phoenix/requirements.txt
requests>=2.32.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pyperclip>=1.9.0
pywinauto>=0.6.8
PyNaCl>=1.5.0
